Rename new RankSVM code files. Remove '_2'
